+++ src/guillo-gmi/views/item.js
@@ -4,11 +4,13 @@ import { useTraversal } from '../contexts'
 import { PanelPermissions } from '../components/panel/permissions'
 import { PanelBehaviors } from '../components/panel/behaviors'
 import { PanelProperties } from '../components/panel/properties'
+import { PanelActions } from '../components/panel/actions'
 
 const tabs = {
   Properties: PanelProperties,
   Behaviors: PanelBehaviors,
   Permissions: PanelPermissions,
+  Actions: PanelActions,
 }
 
 const tabsPermissions = {